requirements.txt
asgiref==3.3.1
beautifulsoup4==4.9.3
bootstrap4==0.1.0
Django==3.1.6
django-bootstrap4==2.3.1
django-chartjs==2.2.1
django-extensions==3.1.3
django-stdimage==5.1.1
gunicorn==20.0.4
mysql==0.0.2
mysqlclient==2.0.3
Pillow==8.1.0
progressbar2==3.53.1
psycopg2-binary==2.8.6
PyMySQL==1.0.2
python-utils==2.5.6
pytz==2021.1
six==1.15.0
soupsieve==2.2
sqlparse==0.4.1
whitenoise==5.2.0
